+++ src/main/java/br/com/springweb/repositorys/ClienteRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import br.com.springweb.entities.Cliente;
 
 @Repository
-public interface ClienteRepository extends JpaRepository<Cliente,Long> {
+public interface ClienteRepository extends JpaRepository<Cliente,Integer> {
 
 }
