+++ src/main/java/br/com/springweb/repositorys/EnderecoRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import br.com.springweb.entities.Endereco;
 
 @Repository
-public interface EnderecoRepository extends JpaRepository<Endereco,Long>{
+public interface EnderecoRepository extends JpaRepository<Endereco,Integer>{
 
 }
